+++ client/src/features/relatedLists/RelatedListAccordionItem.tsx
@@ -1,5 +1,4 @@
 import React, { useState } from 'react'
-import styled from 'styled-components'
 import { isNull } from 'lodash'
 
 import { pushClientEvent } from '../../lib/pushClientEvent'
@@ -7,6 +6,7 @@ import { useGetRelatedListsQuery } from '../../redux/api/ml_api'
 import { IHalLink } from '../../types/IHalLink'
 import StyledHr from '../../styles/shared/Hr'
 import theme from '../../styles/theme'
+import StyledAccordionButton from '../../styles/features/relatedLists/AccordionButton'
 
 import RelatedList from './RelatedList'
 
@@ -15,15 +15,6 @@ interface IProps {
   halLink: string
   index: number
 }
-
-const StyledAccordionButton = styled.button`
-  color: #222222;
-  letter-spacing: 0;
-  text-align: left;
-  line-height: 56px;
-  font-weight: 200;
-  font-size: 1.5rem;
-`
 
 /**
  * Renders the accordion item containing the accordion header and retrieves the HAL link data
@@ -111,7 +102,10 @@ const RelatedListAccordionItem: React.FC<IProps> = ({
           </div>
         </div>
       </div>
-      <StyledHr className="w-100" color={theme.color.lightGray} />
+      <StyledHr
+        className="w-100 relatedListAccordionItemHr"
+        color={theme.color.lightGray}
+      />
     </React.Fragment>
   )
 }