+++ client/src/features/set/About.tsx
@@ -62,7 +62,7 @@ const About: React.FC<IObject> = ({ data }) => {
         )}
         <TextContainer label="Work Class" textLabelClassName="col-md-3">
           <TextValue values={[entityClass]} className="col-md-9" />
-          <StyledHr />
+          <StyledHr className="setClassHr" />
         </TextContainer>
         {identifiers.length > 0 && (
           <IdentifiersContainer identifiers={identifiers} id="sets" />