+++ client/src/styles/features/results/ResultsHeader.ts
@@ -1,8 +1,10 @@
 import styled from 'styled-components'
 
+import theme from '../../theme'
+
 const ResultsHeader = styled.h2`
   font-size: 2em;
-  color: #222222;
+  color: ${theme.color.black};
   letter-spacing: 0;
   text-align: left;
   font-weight: 200;