+++ client/CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- Changed mobile layout of place page ([#432](https://github.com/project-lux/lux-frontend/issues/432)).
+
 ### Removed
 
 ### Fixed